All Doctor Pages - API config bugs
Here is the *Actual* commit for all Doctor Pages after separating and updating for clarity including:

Doctor
        - Landing/home page doctor_home.py
        - Alerts for all alerts doctor_alerts.py
        - Inbox for all messages doctor_inbox.py
        - Patients for all patient portals doctor_patients.py

Note for the team: if you run this from the port you will notice an issue with the API connection for vitals - i think we might need to add a PatientID for Vitals or some pointer to patients since there seems to be some sort of issue there. I will investigate in a separate commit!